+++ core/api-server/tests/cron.js
@@ -1,5 +1,6 @@
 const { expect } = require('chai');
 const querystring = require('querystring');
+const HttpStatus = require('http-status-codes');
 const { pipelines, workerStub } = require('./mocks');
 const { request } = require('./utils');
 let restUrl, baseUrl;
@@ -14,23 +15,13 @@ describe('Cron', () => {
         before(() => {
             restPath = `${restUrl}/cron/results`;
         });
-        it('should throw Method Not Allowed', async () => {
-            const options = {
-                uri: restPath,
-                body: {}
-            };
-            const response = await request(options);
-            expect(response.body).to.have.property('error');
-            expect(response.body.error.code).to.equal(405);
-            expect(response.body.error.message).to.equal('Method Not Allowed');
-        });
         it('should throw status Not Found with params', async () => {
             const options = {
                 uri: restPath + '/no_such_id',
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(404);
+            expect(response.body.error.code).to.equal(HttpStatus.NOT_FOUND);
             expect(response.body.error.message).to.equal('cron results no_such_id Not Found');
         });
         it('should throw validation error of required property name', async () => {
@@ -39,7 +30,7 @@ describe('Cron', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal("data should have required property 'name'");
         });
         it('should throw validation error of order property', async () => {
@@ -49,7 +40,7 @@ describe('Cron', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.contain("data.order should be equal to one of the allowed values");
         });
         it('should throw validation error of sort property', async () => {
@@ -59,7 +50,7 @@ describe('Cron', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.contain("data.sort should be equal to one of the allowed values");
         });
         it('should throw validation error of limit should be >= 1', async () => {
@@ -69,7 +60,7 @@ describe('Cron', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal("data.limit should be >= 1");
         });
         it('should throw validation error of limit should be integer', async () => {
@@ -79,7 +70,7 @@ describe('Cron', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal("data.limit should be integer");
         });
         it('should succeed to get cron results', async () => {
@@ -101,7 +92,7 @@ describe('Cron', () => {
             };
             const response = await request(options);
             const result = response.body.map(r => r.data).sort();
-            expect(response.response.statusCode).to.equal(200);
+            expect(response.response.statusCode).to.equal(HttpStatus.OK);
             expect(result).to.deep.equal(data);
             expect(response.body[0]).to.have.property('jobId');
             expect(response.body[0]).to.have.property('data');
@@ -116,23 +107,13 @@ describe('Cron', () => {
         before(() => {
             restPath = `${restUrl}/cron/status`;
         });
-        it('should throw Method Not Allowed', async () => {
-            const options = {
-                uri: restPath,
-                body: {}
-            };
-            const response = await request(options);
-            expect(response.body).to.have.property('error');
-            expect(response.body.error.code).to.equal(405);
-            expect(response.body.error.message).to.equal('Method Not Allowed');
-        });
         it('should throw status Not Found with params', async () => {
             const options = {
                 uri: restPath + '/no_such_id',
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(404);
+            expect(response.body.error.code).to.equal(HttpStatus.NOT_FOUND);
             expect(response.body.error.message).to.equal('cron status no_such_id Not Found');
         });
         it('should throw validation error of required property name', async () => {
@@ -141,7 +122,7 @@ describe('Cron', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal("data should have required property 'name'");
         });
         it('should throw validation error of order property', async () => {
@@ -151,7 +132,7 @@ describe('Cron', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.contain("data.order should be equal to one of the allowed values");
         });
         it('should throw validation error of sort property', async () => {
@@ -161,7 +142,7 @@ describe('Cron', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.contain("data.sort should be equal to one of the allowed values");
         });
         it('should throw validation error of limit should be >= 1', async () => {
@@ -171,7 +152,7 @@ describe('Cron', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal("data.limit should be >= 1");
         });
         it('should throw validation error of limit should be integer', async () => {
@@ -181,7 +162,7 @@ describe('Cron', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal("data.limit should be integer");
         });
         it('should succeed to get cron status', async () => {
@@ -201,7 +182,7 @@ describe('Cron', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.response.statusCode).to.equal(200);
+            expect(response.response.statusCode).to.equal(HttpStatus.OK);
             expect(response.body[0]).to.have.property('jobId');
             expect(response.body[0]).to.have.property('level');
             expect(response.body[0]).to.have.property('pipeline');
@@ -214,17 +195,6 @@ describe('Cron', () => {
         before(() => {
             restPath = `${restUrl}/cron/start`;
         });
-        it('should throw Method Not Allowed', async () => {
-            const options = {
-                method: 'GET',
-                uri: restPath,
-                body: {}
-            };
-            const response = await request(options);
-            expect(response.body).to.have.property('error');
-            expect(response.body.error.code).to.equal(405);
-            expect(response.body.error.message).to.equal('Method Not Allowed');
-        });
         it('should throw status Not Found with params', async () => {
             const options = {
                 uri: restPath,
@@ -233,7 +203,7 @@ describe('Cron', () => {
                 }
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(404);
+            expect(response.body.error.code).to.equal(HttpStatus.NOT_FOUND);
             expect(response.body.error.message).to.equal('pipeline no_such_name Not Found');
         });
         it('should throw validation error of required property name', async () => {
@@ -241,7 +211,7 @@ describe('Cron', () => {
                 uri: restPath
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal(`data should have required property 'name'`);
         });
         it('should throw validation error of invalid cron', async () => {
@@ -253,7 +223,7 @@ describe('Cron', () => {
                 }
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal(`data.pattern should match format "cron"`);
         });
         it('should success to start cron', async () => {
@@ -301,17 +271,6 @@ describe('Cron', () => {
         before(() => {
             restPath = `${restUrl}/cron/stop`;
         });
-        it('should throw Method Not Allowed', async () => {
-            const options = {
-                method: 'GET',
-                uri: restPath,
-                body: {}
-            };
-            const response = await request(options);
-            expect(response.body).to.have.property('error');
-            expect(response.body.error.code).to.equal(405);
-            expect(response.body.error.message).to.equal('Method Not Allowed');
-        });
         it('should throw status Not Found with params', async () => {
             const options = {
                 uri: restPath,
@@ -320,7 +279,7 @@ describe('Cron', () => {
                 }
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(404);
+            expect(response.body.error.code).to.equal(HttpStatus.NOT_FOUND);
             expect(response.body.error.message).to.equal('pipeline no_such_name Not Found');
         });
         it('should throw validation error of required property name', async () => {
@@ -329,7 +288,7 @@ describe('Cron', () => {
                 method
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal(`data should have required property 'name'`);
         });
         it('should success to stop cron', async () => {
