+++ core/api-server/tests/exec-raw.js
@@ -1,0 +1,245 @@
+const { expect } = require('chai');
+const HttpStatus = require('http-status-codes');
+const { request } = require('./utils');
+let restUrl;
+
+describe('Executions', () => {
+    before(() => {
+        restUrl = global.testParams.restUrl;
+    });
+    describe('/exec/raw', () => {
+        let restPath = null;
+        before(() => {
+            restPath = `${restUrl}/exec/raw`;
+        });
+        it('should throw validation error of required property name', async () => {
+            const options = {
+                uri: restPath,
+                body: {}
+            };
+            const response = await request(options);
+            expect(response.body).to.have.property('error');
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
+            expect(response.body.error.message).to.equal("data should have required property 'name'");
+        });
+        it('should throw validation error of data.name should be string', async () => {
+            const options = {
+                uri: restPath,
+                body: {
+                    name: {}
+                }
+            };
+            const response = await request(options);
+            expect(response.body).to.have.property('error');
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
+            expect(response.body.error.message).to.equal('data.name should be string');
+        });
+        it('should throw validation error of name should NOT be shorter than 1 characters"', async () => {
+            const options = {
+                uri: restPath,
+                body: {
+                    name: ''
+                }
+            };
+            const response = await request(options);
+            expect(response.body).to.have.property('error');
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
+            expect(response.body.error.message).to.equal('data.name should NOT be shorter than 1 characters');
+        });
+        it('should throw validation error of required property nodes', async () => {
+            const options = {
+                uri: restPath,
+                body: {
+                    name: 'string'
+                }
+            };
+            const response = await request(options);
+            expect(response.body).to.have.property('error');
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
+            expect(response.body.error.message).to.equal("data should have required property 'nodes'");
+        });
+        it('should throw validation error of required property nodes.nodeName', async () => {
+            const options = {
+                uri: restPath,
+                body: {
+                    name: 'string',
+                    nodes: [
+                        {
+                            algorithmName: 'green-alg',
+                            input: [
+                                {}
+                            ]
+                        }
+                    ]
+                }
+            };
+            const response = await request(options);
+            expect(response.body).to.have.property('error');
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
+            expect(response.body.error.message).to.equal("data.nodes[0] should have required property 'nodeName'");
+        });
+        it('should throw validation error of required property nodes.algorithmName', async () => {
+            const options = {
+                uri: restPath,
+                body: {
+                    name: 'string',
+                    nodes: [
+                        {
+                            nodeName: 'string',
+                            input: [
+                                {}
+                            ]
+                        }
+                    ]
+                }
+            };
+            const response = await request(options);
+            expect(response.body).to.have.property('error');
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
+            expect(response.body.error.message).to.equal("data.nodes[0] should have required property 'algorithmName'");
+        });
+        it('should throw validation error of nodes.input should be array', async () => {
+            const options = {
+                uri: restPath,
+                body: {
+                    name: 'string',
+                    nodes: [
+                        {
+                            nodeName: 'string',
+                            algorithmName: 'green-alg',
+                            input: null
+                        }
+                    ]
+                }
+            };
+            const response = await request(options);
+            expect(response.body).to.have.property('error');
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
+        });
+        it('should throw validation error of data should NOT have additional properties', async () => {
+            const options = {
+                uri: restPath,
+                body: {
+                    name: 'string',
+                    nodes: [
+                        {
+                            nodeName: 'string',
+                            algorithmName: 'green-alg',
+                            input: []
+                        }
+                    ],
+                    additionalProps: {
+                        bla: 60,
+                        blabla: 'info'
+                    }
+                }
+            };
+            const response = await request(options);
+            expect(response.body).to.have.property('error');
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
+            expect(response.body.error.message).to.equal('data should NOT have additional properties');
+        });
+        it('should throw validation error of duplicate node', async () => {
+            const options = {
+                uri: restPath,
+                body: {
+                    name: 'string',
+                    nodes: [
+                        {
+                            nodeName: 'dup',
+                            algorithmName: 'green-alg',
+                            input: []
+                        },
+                        {
+                            nodeName: 'dup',
+                            algorithmName: 'green-alg',
+                            input: []
+                        }
+                    ]
+                }
+            };
+            const response = await request(options);
+            expect(response.body).to.have.property('error');
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
+            expect(response.body.error.message).to.equal('found duplicate node dup');
+        });
+        it('should throw validation error priority range', async () => {
+            const options = {
+                uri: restPath,
+                body: {
+                    name: 'string',
+                    nodes: [
+                        {
+                            nodeName: 'dup',
+                            algorithmName: 'green-alg',
+                            input: []
+                        }
+                    ],
+                    priority: 8
+                }
+            };
+            const response = await request(options);
+            expect(response.body).to.have.property('error');
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
+            expect(response.body.error.message).to.equal('data.priority should be <= 5');
+        });
+        const invalidStartAndEndChars = ['/', '*', '#', '"', '%'];
+        invalidStartAndEndChars.forEach((v) => {
+            it(`should throw validation error pipeline name include ${v}`, async () => {
+                const options = {
+                    uri: restPath,
+                    body: {
+                        name: `exec${v}raw`,
+                        nodes: [
+                            {
+                                nodeName: 'string',
+                                algorithmName: 'green-alg',
+                                input: []
+                            }
+                        ]
+                    }
+                };
+                const response = await request(options);
+                expect(response.body).to.have.property('error');
+                expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
+                expect(response.body.error.message).to.equal('pipeline name must contain only alphanumeric, dash, dot or underscore');
+            });
+            it(`should throw validation error pipeline name start with ${v}`, async () => {
+                const options = {
+                    uri: restPath,
+                    body: {
+                        name: `${v}xecraw`,
+                        nodes: [
+                            {
+                                nodeName: 'string',
+                                algorithmName: 'green-alg',
+                                input: []
+                            }
+                        ]
+                    }
+                };
+                const response = await request(options);
+                expect(response.body).to.have.property('error');
+                expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
+                expect(response.body.error.message).to.equal('pipeline name must contain only alphanumeric, dash, dot or underscore');
+            });
+        });
+        it('should succeed and return job id', async () => {
+            const options = {
+                uri: restPath,
+                body: {
+                    name: 'exec_raw',
+                    nodes: [
+                        {
+                            nodeName: 'string',
+                            algorithmName: 'green-alg',
+                            input: []
+                        }
+                    ]
+                }
+            };
+            const response = await request(options);
+            expect(response.body).to.have.property('jobId');
+        });
+    });
+});
