+++ core/api-server/tests/exec-stored.js
@@ -1,0 +1,103 @@
+const { expect } = require('chai');
+const HttpStatus = require('http-status-codes');
+const { request } = require('./utils');
+let restUrl;
+
+describe('Executions', () => {
+    before(() => {
+        restUrl = global.testParams.restUrl;
+    });
+    describe('/exec/stored', () => {
+        let restPath = null;
+        before(() => {
+            restPath = `${restUrl}/exec/stored`;
+        });
+        it('should throw validation error of required property name', async () => {
+            const options = {
+                uri: restPath,
+                body: {}
+            };
+            const response = await request(options);
+            expect(response.body).to.have.property('error');
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
+            expect(response.body.error.message).to.equal("data should have required property 'name'");
+        });
+        it('should throw validation error of data.name should be string', async () => {
+            const options = {
+                uri: restPath,
+                body: {
+                    name: {}
+                }
+            };
+            const response = await request(options);
+            expect(response.body).to.have.property('error');
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
+            expect(response.body.error.message).to.equal('data.name should be string');
+        });
+        it('should throw validation error of name should NOT be shorter than 1 characters"', async () => {
+            const options = {
+                uri: restPath,
+                body: {
+                    name: ''
+                }
+            };
+            const response = await request(options);
+            expect(response.body).to.have.property('error');
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
+            expect(response.body.error.message).to.equal('data.name should NOT be shorter than 1 characters');
+        });
+        it('should throw validation error of data should NOT have additional properties', async () => {
+            const options = {
+                uri: restPath,
+                body: {
+                    name: 'string',
+                    nodes: [
+                        {
+                            nodeName: 'string',
+                            algorithmName: 'green-alg',
+                            input: []
+                        }
+                    ]
+                }
+            };
+            const response = await request(options);
+            expect(response.body).to.have.property('error');
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
+            expect(response.body.error.message).to.equal('data should NOT have additional properties');
+        });
+        it('should throw pipeline not found', async () => {
+            const options = {
+                uri: restPath,
+                body: {
+                    name: 'not_found'
+                }
+            };
+            const response = await request(options);
+            expect(response.body).to.have.property('error');
+            expect(response.body.error.code).to.equal(HttpStatus.NOT_FOUND);
+            expect(response.body.error.message).to.equal('pipeline not_found Not Found');
+        });
+        it('should throw validation error pipeline name', async () => {
+            const options = {
+                uri: restPath,
+                body: {
+                    name: 'exec/stored'
+                }
+            };
+            const response = await request(options);
+            expect(response.body).to.have.property('error');
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
+            expect(response.body.error.message).to.equal('pipeline name must contain only alphanumeric, dash, dot or underscore');
+        });
+        it('should succeed and return job id', async () => {
+            const options = {
+                uri: restPath,
+                body: {
+                    name: 'flow1'
+                }
+            };
+            const response = await request(options);
+            expect(response.body).to.have.property('jobId');
+        });
+    });
+});
