+++ core/api-server/tests/pipelines.js
@@ -1,5 +1,6 @@
 const { expect } = require('chai');
 const querystring = require('querystring');
+const HttpStatus = require('http-status-codes');
 const { workerStub } = require('./mocks');
 const { request } = require('./utils');
 let restUrl;
@@ -13,23 +14,13 @@ describe('Pipelines', () => {
         before(() => {
             restPath = `${restUrl}/pipelines/results/stored`;
         });
-        it('should throw Method Not Allowed', async () => {
-            const options = {
-                uri: restPath,
-                body: {}
-            };
-            const response = await request(options);
-            expect(response.body).to.have.property('error');
-            expect(response.body.error.code).to.equal(405);
-            expect(response.body.error.message).to.equal('Method Not Allowed');
-        });
         it('should throw status Not Found with params', async () => {
             const options = {
                 uri: restPath + '/no_such_id',
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(404);
+            expect(response.body.error.code).to.equal(HttpStatus.NOT_FOUND);
             expect(response.body.error.message).to.equal('pipeline results no_such_id Not Found');
         });
         it('should throw validation error of required property name', async () => {
@@ -38,7 +29,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal("data should have required property 'name'");
         });
         it('should throw validation error of order property', async () => {
@@ -48,7 +39,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.contain("data.order should be equal to one of the allowed values");
         });
         it('should throw validation error of sort property', async () => {
@@ -58,7 +49,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.contain("data.sort should be equal to one of the allowed values");
         });
         it('should throw validation error of limit should be >= 1', async () => {
@@ -68,7 +59,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal("data.limit should be >= 1");
         });
         it('should throw validation error of limit should be integer', async () => {
@@ -78,7 +69,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal("data.limit should be integer");
         });
         it('should succeed to get pipelines results', async () => {
@@ -100,7 +91,7 @@ describe('Pipelines', () => {
             };
             const response = await request(options);
             const result = response.body.map(r => r.data).sort();
-            expect(response.response.statusCode).to.equal(200);
+            expect(response.response.statusCode).to.equal(HttpStatus.OK);
             expect(result).to.deep.equal(data);
             expect(response.body[0]).to.have.property('jobId');
             expect(response.body[0]).to.have.property('data');
@@ -115,23 +106,13 @@ describe('Pipelines', () => {
         before(() => {
             restPath = `${restUrl}/pipelines/results/raw`;
         });
-        it('should throw Method Not Allowed', async () => {
-            const options = {
-                uri: restPath,
-                body: {}
-            };
-            const response = await request(options);
-            expect(response.body).to.have.property('error');
-            expect(response.body.error.code).to.equal(405);
-            expect(response.body.error.message).to.equal('Method Not Allowed');
-        });
         it('should throw status Not Found with params', async () => {
             const options = {
                 uri: restPath + '/no_such_id',
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(404);
+            expect(response.body.error.code).to.equal(HttpStatus.NOT_FOUND);
             expect(response.body.error.message).to.equal('pipeline results no_such_id Not Found');
         });
         it('should throw validation error of required property name', async () => {
@@ -140,7 +121,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal("data should have required property 'name'");
         });
         it('should throw validation error of order property', async () => {
@@ -150,7 +131,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.contain("data.order should be equal to one of the allowed values");
         });
         it('should throw validation error of sort property', async () => {
@@ -160,7 +141,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.contain("data.sort should be equal to one of the allowed values");
         });
         it('should throw validation error of limit should be >= 1', async () => {
@@ -170,7 +151,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal("data.limit should be >= 1");
         });
         it('should throw validation error of limit should be integer', async () => {
@@ -180,7 +161,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal("data.limit should be integer");
         });
         it('should succeed to get pipelines results', async () => {
@@ -209,7 +190,7 @@ describe('Pipelines', () => {
             };
             const response = await request(options);
             const result = response.body.map(r => r.data).sort();
-            expect(response.response.statusCode).to.equal(200);
+            expect(response.response.statusCode).to.equal(HttpStatus.OK);
             expect(result).to.deep.equal(data);
             expect(response.body[0]).to.have.property('jobId');
             expect(response.body[0]).to.have.property('data');
@@ -224,23 +205,14 @@ describe('Pipelines', () => {
         before(() => {
             restPath = `${restUrl}/pipelines/status/stored`;
         });
-        it('should throw Method Not Allowed', async () => {
-            const options = {
-                uri: restPath,
-                body: {}
-            };
-            const response = await request(options);
-            expect(response.body).to.have.property('error');
-            expect(response.body.error.code).to.equal(405);
-            expect(response.body.error.message).to.equal('Method Not Allowed');
-        });
+
         it('should throw status Not Found with params', async () => {
             const options = {
                 uri: restPath + '/no_such_id',
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(404);
+            expect(response.body.error.code).to.equal(HttpStatus.NOT_FOUND);
             expect(response.body.error.message).to.equal('pipeline status no_such_id Not Found');
         });
         it('should throw validation error of required property name', async () => {
@@ -249,7 +221,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal("data should have required property 'name'");
         });
         it('should throw validation error of order property', async () => {
@@ -259,7 +231,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.contain("data.order should be equal to one of the allowed values");
         });
         it('should throw validation error of sort property', async () => {
@@ -269,7 +241,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.contain("data.sort should be equal to one of the allowed values");
         });
         it('should throw validation error of limit should be >= 1', async () => {
@@ -279,7 +251,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal("data.limit should be >= 1");
         });
         it('should throw validation error of limit should be integer', async () => {
@@ -289,7 +261,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal("data.limit should be integer");
         });
         it('should succeed to get pipelines status', async () => {
@@ -309,7 +281,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.response.statusCode).to.equal(200);
+            expect(response.response.statusCode).to.equal(HttpStatus.OK);
             expect(response.body).to.have.lengthOf(limit)
             expect(response.body[0]).to.have.property('jobId');
             expect(response.body[0]).to.have.property('level');
@@ -323,24 +295,13 @@ describe('Pipelines', () => {
         before(() => {
             restPath = `${restUrl}/pipelines/status/raw`;
         });
-        it('should throw Method Not Allowed', async () => {
-            const options = {
-
-                uri: restPath,
-                body: {}
-            };
-            const response = await request(options);
-            expect(response.body).to.have.property('error');
-            expect(response.body.error.code).to.equal(405);
-            expect(response.body.error.message).to.equal('Method Not Allowed');
-        });
         it('should throw status Not Found with params', async () => {
             const options = {
                 uri: restPath + '/no_such_id',
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(404);
+            expect(response.body.error.code).to.equal(HttpStatus.NOT_FOUND);
             expect(response.body.error.message).to.equal('pipeline status no_such_id Not Found');
         });
         it('should throw validation error of required property name', async () => {
@@ -349,7 +310,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal("data should have required property 'name'");
         });
         it('should throw validation error of order property', async () => {
@@ -359,7 +320,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.contain("data.order should be equal to one of the allowed values");
         });
         it('should throw validation error of sort property', async () => {
@@ -369,7 +330,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.contain("data.sort should be equal to one of the allowed values");
         });
         it('should throw validation error of limit should be >= 1', async () => {
@@ -379,7 +340,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal("data.limit should be >= 1");
         });
         it('should throw validation error of limit should be integer', async () => {
@@ -389,7 +350,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal("data.limit should be integer");
         });
         it('should succeed to get pipelines statuses', async () => {
@@ -416,7 +377,7 @@ describe('Pipelines', () => {
                 method: 'GET'
             };
             const response = await request(options);
-            expect(response.response.statusCode).to.equal(200);
+            expect(response.response.statusCode).to.equal(HttpStatus.OK);
             expect(response.body).to.have.lengthOf(limit)
             expect(response.body[0]).to.have.property('jobId');
             expect(response.body[0]).to.have.property('level');
