+++ core/api-server/tests/readme.js
@@ -1,5 +1,6 @@
 const { expect } = require('chai');
 const fse = require('fs-extra');
+const HttpStatus = require('http-status-codes');
 const { algorithms, pipelines } = require('./mocks');
 const { request } = require('./utils');
 let restUrl;
@@ -20,7 +21,7 @@ describe('ReadMe', () => {
             };
             const response = await request(options);
             expect(response.body).to.have.property('error');
-            expect(response.body.error.code).to.equal(404);
+            expect(response.body.error.code).to.equal(HttpStatus.NOT_FOUND);
             expect(response.body.error.message).to.equal("pipeline not_exist Not Found");
         });
         it('should throw validation error of readme Not Found', async () => {
@@ -30,7 +31,7 @@ describe('ReadMe', () => {
             };
             const response = await request(options);
             expect(response.body).to.have.property('error');
-            expect(response.body.error.code).to.equal(404);
+            expect(response.body.error.code).to.equal(HttpStatus.NOT_FOUND);
             expect(response.body.error.message).to.equal(`readme ${pipelines[0].name} Not Found`);
         });
         it('should success to insert readme', async () => {
@@ -45,7 +46,7 @@ describe('ReadMe', () => {
             };
             const response1 = await request(options1);
             expect(response1.body).to.have.property('path');
-            expect(response1.response.statusCode).to.equal(201);
+            expect(response1.response.statusCode).to.equal(HttpStatus.CREATED);
 
             const options2 = {
                 uri,
@@ -70,7 +71,7 @@ describe('ReadMe', () => {
             };
             const response1 = await request(insert);
             expect(response1.body).to.have.property('path');
-            expect(response1.response.statusCode).to.equal(201);
+            expect(response1.response.statusCode).to.equal(HttpStatus.CREATED);
 
             const put = {
                 uri,
@@ -100,7 +101,7 @@ describe('ReadMe', () => {
             };
             const response1 = await request(insert);
             expect(response1.body).to.have.property('path');
-            expect(response1.response.statusCode).to.equal(201);
+            expect(response1.response.statusCode).to.equal(HttpStatus.CREATED);
 
             const options2 = {
                 uri,
@@ -116,7 +117,7 @@ describe('ReadMe', () => {
             };
             const response3 = await request(get);
             expect(response3.body).to.have.property('error');
-            expect(response3.body.error.code).to.equal(404);
+            expect(response3.body.error.code).to.equal(HttpStatus.NOT_FOUND);
             expect(response3.body.error.message).to.equal(`readme ${pipelines[0].name} Not Found`);
 
         });
@@ -133,7 +134,7 @@ describe('ReadMe', () => {
             };
             const response = await request(options);
             expect(response.body).to.have.property('error');
-            expect(response.body.error.code).to.equal(404);
+            expect(response.body.error.code).to.equal(HttpStatus.NOT_FOUND);
             expect(response.body.error.message).to.equal("algorithm not_exist Not Found");
         });
         it('should throw validation error of readme Not Found', async () => {
@@ -143,7 +144,7 @@ describe('ReadMe', () => {
             };
             const response = await request(options);
             expect(response.body).to.have.property('error');
-            expect(response.body.error.code).to.equal(404);
+            expect(response.body.error.code).to.equal(HttpStatus.NOT_FOUND);
             expect(response.body.error.message).to.equal(`readme ${algorithms[0].name} Not Found`);
         });
         it('should success to insert readme', async () => {
@@ -159,7 +160,7 @@ describe('ReadMe', () => {
             };
             const response1 = await request(insert);
             expect(response1.body).to.have.property('path');
-            expect(response1.response.statusCode).to.equal(201);
+            expect(response1.response.statusCode).to.equal(HttpStatus.CREATED);
 
             const options2 = {
                 uri,
@@ -184,7 +185,7 @@ describe('ReadMe', () => {
             };
             const response1 = await request(insert);
             expect(response1.body).to.have.property('path');
-            expect(response1.response.statusCode).to.equal(201);
+            expect(response1.response.statusCode).to.equal(HttpStatus.CREATED);
 
             const put = {
                 uri,
@@ -214,7 +215,7 @@ describe('ReadMe', () => {
             };
             const response1 = await request(insert);
             expect(response1.body).to.have.property('path');
-            expect(response1.response.statusCode).to.equal(201);
+            expect(response1.response.statusCode).to.equal(HttpStatus.CREATED);
 
             const options2 = {
                 uri,
@@ -230,7 +231,7 @@ describe('ReadMe', () => {
             };
             const response3 = await request(get);
             expect(response3.body).to.have.property('error');
-            expect(response3.body.error.code).to.equal(404);
+            expect(response3.body.error.code).to.equal(HttpStatus.NOT_FOUND);
             expect(response3.body.error.message).to.equal(`readme ${algorithms[0].name} Not Found`);
 
         });
