+++ core/api-server/tests/webhooks.js
@@ -1,4 +1,5 @@
 const { expect } = require('chai');
+const HttpStatus = require('http-status-codes');
 const storageManager = require('@hkube/storage-manager');
 const stateManager = require('../lib/state/state-manager');
 const { webhookStub } = require('./mocks');
@@ -68,7 +69,7 @@ describe('Webhooks', () => {
             };
             const response = await request(options);
             expect(response.body).to.have.property('error');
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal('data.webhooks.result should match format "url"');
         });
         it('should throw webhooks validation error of NOT have additional properties', async () => {
@@ -90,7 +91,7 @@ describe('Webhooks', () => {
             };
             const response = await request(options);
             expect(response.body).to.have.property('error');
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal('data.webhooks should NOT have additional properties');
         });
         it('should succeed to store pipeline with webhooks', async () => {
@@ -119,7 +120,7 @@ describe('Webhooks', () => {
                 uri: `${restUrl}/webhooks/results/no_such`
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(404);
+            expect(response.body.error.code).to.equal(HttpStatus.NOT_FOUND);
             expect(response.body.error.message).to.equal('webhook no_such Not Found');
         });
         it('should succeed to send webhook and get results', async () => {
@@ -202,7 +203,7 @@ describe('Webhooks', () => {
             };
             const response = await request(options);
             expect(response.body).to.have.property('error');
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal('data.webhooks.progress should match format "url"');
         });
         it('should throw webhooks validation error of NOT have additional properties', async () => {
@@ -224,7 +225,7 @@ describe('Webhooks', () => {
             };
             const response = await request(options);
             expect(response.body).to.have.property('error');
-            expect(response.body.error.code).to.equal(400);
+            expect(response.body.error.code).to.equal(HttpStatus.BAD_REQUEST);
             expect(response.body.error.message).to.equal('data.webhooks should NOT have additional properties');
         });
         it('should succeed to store pipeline with webhooks', async () => {
@@ -253,7 +254,7 @@ describe('Webhooks', () => {
                 uri: `${restUrl}/webhooks/status/no_such`
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(404);
+            expect(response.body.error.code).to.equal(HttpStatus.NOT_FOUND);
             expect(response.body.error.message).to.equal('webhook no_such Not Found');
         });
         it('should throw error when no such jobId', async function () {
@@ -262,7 +263,7 @@ describe('Webhooks', () => {
                 uri: `${restUrl}/webhooks/no_such`
             };
             const response = await request(options);
-            expect(response.body.error.code).to.equal(404);
+            expect(response.body.error.code).to.equal(HttpStatus.NOT_FOUND);
             expect(response.body.error.message).to.equal('webhook no_such Not Found');
         });
         it('should succeed and return webhooks progress', async function () {
